backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Flask-SocketIO==5.5.1
PyJWT==2.10.1
pymongo==4.14.1
bcrypt==4.3.0
python-dotenv==1.1.1
gunicorn==23.0.0
python-socketio==5.12.1
eventlet==0.39.0
